Due to the 2-color chemistry that it is based on, the NovaSeq sequencer (presumably also NextSeq) will read beyond the end of a molecule (asfter having read the adapter) and then produce high quality "G" bases when its it has fallen off. This can be aligned (for example to position chr2:32916229 in hg38) albeit with mapping quality 0 (generally).
This PR modified some of the metric collection tools to observe MQ0 adapter reads (previously only unmapped reads would be considered as possible candidates for adapter reads) and account for them as such.

It also takes the opportunity to clean up some of the metrics files (extracted WgsMetrics into it's own file and refactored the various flavors of TargetMetrics so that they all extend a base class)
