+++ src/main/java/picard/analysis/CollectRawWgsMetrics.java
@@ -122,6 +122,7 @@ public class CollectRawWgsMetrics extends CollectWgsMetrics{
         public RawWgsMetrics(final IntervalList intervals,
                              final Histogram<Integer> highQualityDepthHistogram,
                              final Histogram<Integer> unfilteredDepthHistogram,
+                             final double pctExcludedByAdapter,
                              final double pctExcludedByMapq,
                              final double pctExcludedByDupes,
                              final double pctExcludedByPairing,
@@ -132,7 +133,7 @@ public class CollectRawWgsMetrics extends CollectWgsMetrics{
                              final int coverageCap,
                              final Histogram<Integer> unfilteredBaseQHistogram,
                              final int sampleSize) {
-            super(intervals, highQualityDepthHistogram, unfilteredDepthHistogram, pctExcludedByMapq, pctExcludedByDupes, pctExcludedByPairing, pctExcludedByBaseq,
+            super(intervals, highQualityDepthHistogram, unfilteredDepthHistogram, pctExcludedByAdapter, pctExcludedByMapq, pctExcludedByDupes, pctExcludedByPairing, pctExcludedByBaseq,
                     pctExcludedByOverlap, pctExcludedByCapping, pctTotal, coverageCap, unfilteredBaseQHistogram, sampleSize);
         }
     }
@@ -141,6 +142,7 @@ public class CollectRawWgsMetrics extends CollectWgsMetrics{
     protected WgsMetrics generateWgsMetrics(final IntervalList intervals,
                                             final Histogram<Integer> highQualityDepthHistogram,
                                             final Histogram<Integer> unfilteredDepthHistogram,
+                                            final double pctExcludedByAdapter,
                                             final double pctExcludedByMapq,
                                             final double pctExcludedByDupes,
                                             final double pctExcludedByPairing,
@@ -155,6 +157,7 @@ public class CollectRawWgsMetrics extends CollectWgsMetrics{
                 intervals,
                 highQualityDepthHistogram,
                 unfilteredDepthHistogram,
+                pctExcludedByAdapter,
                 pctExcludedByMapq,
                 pctExcludedByDupes,
                 pctExcludedByPairing,
