+++ testdata/picard/sam/summary_alignment_stats_test_adapter_reads.sam
@@ -1,0 +1,5 @@
+@HD	VN:1.0	SO:unsorted
+@SQ	SN:chrM	LN:16571
+@RG	ID:62A40.2	SM:RRBS885	LB:whatever	PU:me	PL:ILLUMINA
+HC5CGDSXX190111:1:2512:8730:13197	129	chrM	227	0	49S14M1D88M	=	253	27	AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAAGAGGTTGTGTGGGTGTAGGGGGGGGGCGGGGGGGGGGGAGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG	????????????????????????????????????????????++?+?++'?++???+??+5?+?++++?++'+++?+????????????????????????????????????????????????????????????????????????	MC:Z:69S82M	RG:Z:62A40.2	MQ:i:0	AS:i:85	XT:i:1
+HC5CGDSXX190111:1:2512:8730:13197	65	chrM	253	3	69S82M	=	227	-27	AGATCGGAAGAGCACACGTCTGAACTCCAGTCACGCACATCTATCGCGTATGCCGTCTTCTGCTTGAAAGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG	????????????????????????????????????+????????'???+????????5+5?55?????++????????????????????????????????????????????????????????????????????????????????	MC:Z:49S14M1D88M	RG:Z:62A40.2	MQ:i:0	AS:i:81	XT:i:1
